# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Figranium API",
    "description": "REST API for Figranium — a browser automation and web scraping platform (headless scraping, headful VNC browser sessions, and AI agent execution).",
    "version": "0.14.4"
  },
  "servers": [
    { "url": "http://localhost:11345", "description": "Local server" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after login."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key configured via /api/settings/api-key."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Machine-readable error code." },
          "message": { "type": "string" },
          "details": {},
          "detail": {}
        },
        "required": ["error"]
      },
...
      "ProxyList": {
        "type": "object",
        "description": "Opaque shape returned by listProxies().",
        "additionalProperties": true
      },
...
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Not authenticated",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "BadRequest": {
        "description": "Invalid request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "sessionAuth": [] }, { "apiKeyAuth": [] }],
  "paths": {
...
    "/api/settings/system": {
      "get": {
        "tags": ["Settings"],
        "summary": "Read retention and automatic resource protection status",
        "responses": { "200": { "description": "System settings and runtime status" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Save shared capture and execution retention",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "retentionDays": { "type": ["integer", "null"], "minimum": 1, "maximum": 365 } } } } } },
        "responses": { "200": { "description": "Saved" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
...
    "/api/settings/api-key": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get the configured API key",
        "responses": {
          "200": { "description": "API key", "content": { "application/json": { "schema": { "type": "object", "properties": { "apiKey": { "type": "string", "nullable": true } } } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Set or generate the API key",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "apiKey": { "type": "string", "maxLength": 512 } } } } }
        },
        "responses": {
          "200": { "description": "Saved key", "content": { "application/json": { "schema": { "type": "object", "properties": { "apiKey": { "type": "string" } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/settings/user-agent": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get user-agent configuration",
        "responses": {
          "200": { "description": "User-agent config", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Set user-agent selection",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "selection": { "type": "string", "nullable": true } } } } }
        },
        "responses": {
          "200": { "description": "Updated config", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/settings/gemini-api-key": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get Gemini API keys",
        "responses": {
          "200": { "description": "Keys", "content": { "application/json": { "schema": { "type": "object", "properties": { "geminiApiKeys": { "type": "array", "items": { "type": "string" } } } } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Set Gemini API keys",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "geminiApiKeys": { "type": "array", "items": { "type": "string" } }, "geminiApiKey": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "geminiApiKeys": { "type": "array", "items": { "type": "string" } } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/settings/openai-api-key": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get OpenAI API keys",
        "responses": {
          "200": { "description": "Keys", "content": { "application/json": { "schema": { "type": "object", "properties": { "openAiApiKeys": { "type": "array", "items": { "type": "string" } } } } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Set OpenAI API keys",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "openAiApiKeys": { "type": "array", "items": { "type": "string" } }, "openAiApiKey": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "openAiApiKeys": { "type": "array", "items": { "type": "string" } } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/settings/claude-api-key": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get Claude API keys",
        "responses": {
          "200": { "description": "Keys", "content": { "application/json": { "schema": { "type": "object", "properties": { "claudeApiKeys": { "type": "array", "items": { "type": "string" } } } } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Set Claude API keys",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "claudeApiKeys": { "type": "array", "items": { "type": "string" } }, "claudeApiKey": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "claudeApiKeys": { "type": "array", "items": { "type": "string" } } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/settings/ollama-api-key": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get Ollama endpoints",
        "responses": {
          "200": { "description": "Endpoints", "content": { "application/json": { "schema": { "type": "object", "properties": { "ollamaApiKeys": { "type": "array", "items": { "type": "string" } } } } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Set Ollama endpoints",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "ollamaApiKeys": { "type": "array", "items": { "type": "string" } }, "ollamaApiKey": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "ollamaApiKeys": { "type": "array", "items": { "type": "string" } } } } } } },
          "400": { "description": "Invalid URL", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/settings/proxies": {
      "get": {
        "tags": ["Settings"],
        "summary": "List proxies",
        "responses": {
          "200": { "description": "Proxy list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProxyList" } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Add a proxy",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["server"],
                "properties": {
                  "server": { "type": "string" },
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "label": { "type": "string" },
                  "isRotatingPool": { "type": "boolean" },
                  "estimatedPoolSize": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated proxy list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProxyList" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "tags": ["Settings"],
        "summary": "Bulk delete proxies",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["ids"], "properties": { "ids": { "type": "array", "items": { "type": "string" } } } } } }
        },
        "responses": {
          "200": { "description": "Updated proxy list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProxyList" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/settings/proxies/import": {
      "post": {
        "tags": ["Settings"],
        "summary": "Bulk import proxies",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["proxies"],
                "properties": { "proxies": { "type": "array", "items": { "type": "object", "additionalProperties": true } } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated proxy list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProxyList" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  }
}